64bit tool fixes

+++ neo/tools/radiant/MapInfo.cpp
@@ -95,7 +95,7 @@ BOOL CMapInfo::OnInitDialog()
 
   CMapStringToPtr mapEntity;
 
-  int nValue = 0;
+  intptr_t nValue = 0;
 	for (entity_t* pEntity=entities.next ; pEntity != &entities ; pEntity=pEntity->next)
 	{
     m_nTotalEntities++;
